Simplify class and helper function for primer
